Split DpExtensions/SpExtensions/EmExtensions and add PublicAPI entries
Two follow-ups after rebasing onto current main:

* Repo style requires one public top-level type per file (per
  .github/copilot-instructions.md). The Dp/Sp/Em.cs files each had two
  public types -- the value-class struct plus a static-extensions
  class. Move each `*Extensions` class into its own file.

* Main now wires up Microsoft.CodeAnalysis.PublicApiAnalyzers with
  TreatWarningsAsErrors=true (Directory.Build.props), so every new
  public symbol must be declared in PublicAPI.Unshipped.txt. Add
  entries for Dp, DpExtensions, Em, EmExtensions, FontWeight, Shape,
  Sp, SpExtensions, TextAlign, TextDecoration plus their static
  members, operators, and synthesized record overrides.
