+++ src/ComposeNet.Compose/Dp.cs
@@ -38,16 +38,3 @@ public readonly record struct Dp(float Value)
     /// </summary>
     public static float Pack(Dp? value) => value?.Value ?? 0f;
 }
-
-/// <summary>
-/// Convenience constructors for the <see cref="ComposeNet.Dp"/>
-/// struct: <c>16.Dp()</c>, <c>0.5f.Dp()</c>.
-/// </summary>
-public static class DpExtensions
-{
-    /// <summary>Wrap an integer dp value in a <c>Dp</c>.</summary>
-    public static Dp Dp(this int value) => new(value);
-
-    /// <summary>Wrap a float dp value in a <c>Dp</c>.</summary>
-    public static Dp Dp(this float value) => new(value);
-}
