+++ src/ComposeNet.Compose/Sp.cs
@@ -35,16 +35,3 @@ public readonly record struct Sp(float Value)
         return (1L << 32) | (long)raw;
     }
 }
-
-/// <summary>
-/// Convenience constructors for the <see cref="ComposeNet.Sp"/>
-/// struct: <c>16.Sp()</c>, <c>14.5f.Sp()</c>.
-/// </summary>
-public static class SpExtensions
-{
-    /// <summary>Wrap an integer point size in an <c>Sp</c>.</summary>
-    public static Sp Sp(this int value) => new(value);
-
-    /// <summary>Wrap a float point size in an <c>Sp</c>.</summary>
-    public static Sp Sp(this float value) => new(value);
-}
